% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_DFS_data.R
\name{historical_DFS_data}
\alias{historical_DFS_data}
\title{Historical DFS Data}
\usage{
historical_DFS_data(
  which,
  key,
  tour = "pga",
  site = "draftkings",
  event_id = NULL,
  year = "2023",
  file_format = "json"
)
}
\arguments{
\item{which}{A character string indicating the which of data to retrieve. Possible values are 'ids' and 'points'.}

\item{key}{A character string representing the user's API key for Datagolf.}

\item{tour}{A character string indicating the tour to retrieve data for (e.g. 'pga', 'euro').}

\item{site}{A character string indicating the site to retrieve data for (e.g. 'draftkings', 'fanduel').}

\item{event_id}{A character string representing the ID of the event to retrieve data for.}

\item{year}{A character string representing the year of the event to retrieve data for.}

\item{file_format}{A character string indicating the file format of the response (e.g. 'json', 'csv').}
}
\value{
A data frame containing the requested historical DFS data.
}
\description{
This function returns historical DFS data from the Datagolf API as a dataframe.
}
\examples{
\dontrun{
historical_DFS_data(which = "ids", key = "your_api_key_here")
historical_DFS_data(which = "points", key = "your_api_key_here", tour = "euro", site = "draftkings", event_id = "405", year = "2022", file_format = "csv")
}

}
