% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_df.R
\name{list_to_df}
\alias{list_to_df}
\title{Convert a list of data frames to a single data frame}
\usage{
list_to_df(your_list)
}
\arguments{
\item{your_list}{A list of data frames to combine.}
}
\value{
A single data frame obtained by row binding the data frames in the input list.
}
\description{
This function takes a list of data frames and combines them into a single data frame by row binding.
The function adds any missing columns to each data frame in the list and makes sure that the columns are in the same order.
}
