% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_names.R
\name{switch_names}
\alias{switch_names}
\title{Switch names}
\usage{
switch_names(names)
}
\arguments{
\item{names}{A character vector of names to be switched.}
}
\value{
A character vector of names with the first and last names switched.
}
\description{
This function takes a vector of names and switches the order of the first and last names.
If a name consists of more than two parts (e.g., middle names or initials), the function
will handle those cases as well.
}
\examples{
switch_names(c("John Doe", "Jane A. Smith", "Robert James Williams"))

}
