% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Update data in an RDS file with new data from Rolf package}
\usage{
update_data()
}
\value{
Nothing is returned, but the updated list is saved back to the same RDS file.
}
\description{
This function reads in a list object from an RDS file, and updates it with new
data from the Rolf package. The updated list is saved back to the same RDS file.
}
\details{
This function starts by reading in the list object from the RDS file
using \code{readRDS()}, and loading the \code{Rolf} and \code{dplyr} packages. It then retrieves
new data from the Rolf package using \code{Rolf::historical_raw_data()} and
\code{Rolf::historical_DFS_data()}. The function updates the list object with new
data if the new data is not already in the list, and saves the updated list back
to the same RDS file using \code{saveRDS()}.
}
\examples{
# Run the function to update the data
update_data()

}
\seealso{
\code{\link{readRDS}}, \code{\link{saveRDS}}
}
